serverapp/tg-bot/requirements.txt
pyTelegramBotAPI==4.7.1
PyYAML==6.0
python-docx==0.8.11
requests==2.28.1
